requirements.txt
Flask
flask-cors
pymongo
python-dotenv
flask
pymongo
flask-cors
python-dotenv
